requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-crispy-forms==1.9.2
django-filter==2.3.0
django-mathfilters==1.0.0
html5lib==1.1
idna==2.10
Pillow==7.2.0
PyPDF2==1.26.0
pytz==2020.1
reportlab==3.5.46
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
stripe==2.49.0
urllib3==1.25.10
webencodings==0.5.1
xhtml2pdf==0.2.4
